refactor(filemeta): split filemeta into focused submodules (#1946)
